project/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
ta>=0.7.0
yfinance>=0.1.63

# Machine Learning
statsmodels>=0.12.0
arch>=4.19.0
tensorflow>=2.6.0
keras>=2.6.0

# Data handling
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# Technical Analysis
ta-lib>=0.4.0

# Configuration
pyyaml>=5.4.0
python-json-logger>=2.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.0.0
python-dateutil>=2.8.2

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
sphinx>=4.1.0
sphinx-rtd-theme>=0.5.0